--- front-api/middlewares/ensure_is_admin.ts
@@ -1,24 +1,0 @@
-import type { WorkspaceAwareCtx } from "@front-api/middlewares/ctx";
-import { apiError } from "@front-api/middlewares/utils";
-import { createMiddleware } from "hono/factory";
-
-/**
- * Gate a handler to admin-only access. Must be applied after `workspaceAuth`
- * so `ctx.get("auth")` is set.
- */
-export const ensureIsAdmin = () =>
-  createMiddleware<WorkspaceAwareCtx>(async (ctx, next) => {
-    const auth = ctx.get("auth");
-
-    if (!auth.isAdmin()) {
-      return apiError(ctx, {
-        status_code: 403,
-        api_error: {
-          type: "workspace_auth_error",
-          message: "Only users that are `admins` can perform this action.",
-        },
-      });
-    }
-
-    await next();
-  });
