--- front-api/middlewares/ensure_is_builder.ts
@@ -1,24 +1,0 @@
-import type { WorkspaceAwareCtx } from "@front-api/middlewares/ctx";
-import { apiError } from "@front-api/middlewares/utils";
-import { createMiddleware } from "hono/factory";
-
-/**
- * Gate a handler to builder-or-above access. Must be applied after
- * `workspaceAuth` so `ctx.get("auth")` is set.
- */
-export const ensureIsBuilder = () =>
-  createMiddleware<WorkspaceAwareCtx>(async (ctx, next) => {
-    const auth = ctx.get("auth");
-
-    if (!auth.isBuilder()) {
-      return apiError(ctx, {
-        status_code: 403,
-        api_error: {
-          type: "workspace_auth_error",
-          message: "Only users that are `builders` can perform this action.",
-        },
-      });
-    }
-
-    await next();
-  });
