+++ front-api/routes/w/[wId]/assistant/agent_configurations/batch_update_scope.ts
@@ -1,6 +1,6 @@
 import { updateAgentConfigurationsScope } from "@app/lib/api/assistant/configuration/agent";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsBuilder } from "@front-api/middlewares/ensure_is_builder";
+import { ensureIsBuilder } from "@front-api/middlewares/ensure_role";
 import { apiError } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
