+++ front-api/routes/w/[wId]/credentials/check_bigquery_locations.ts
@@ -1,7 +1,7 @@
 import { CheckBigQueryCredentialsSchema } from "@app/types/oauth/lib";
 import { normalizeError } from "@app/types/shared/utils/error_utils";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsAdmin } from "@front-api/middlewares/ensure_is_admin";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { BigQuery } from "@google-cloud/bigquery";
