+++ front-api/routes/w/[wId]/keys/[id]/index.ts
@@ -7,7 +7,7 @@ import { invalidateKeyCapCache } from "@app/lib/api/programmatic_usage/key_cap";
 import { KeyResource } from "@app/lib/resources/key_resource";
 import type { KeyType } from "@app/types/key";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsAdmin } from "@front-api/middlewares/ensure_is_admin";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
