+++ front-api/routes/w/[wId]/members/index.ts
@@ -2,7 +2,7 @@ import { getMembers } from "@app/lib/api/workspace";
 import type { MembershipsPaginationParams } from "@app/lib/resources/membership_resource";
 import type { UserTypeWithWorkspaces } from "@app/types/user";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsAdmin } from "@front-api/middlewares/ensure_is_admin";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import type { Context } from "hono";
