+++ front-api/routes/w/[wId]/metronome/contract.ts
@@ -8,7 +8,7 @@ import { hasContractSeatSubscription } from "@app/lib/metronome/seats";
 import { isEntreprisePlanPrefix } from "@app/lib/plans/plan_codes";
 import { assertNever } from "@app/types/shared/utils/assert_never";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsAdmin } from "@front-api/middlewares/ensure_is_admin";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
