+++ front-api/routes/w/[wId]/metronome/invoice.ts
@@ -10,7 +10,7 @@ import {
 import type { SupportedCurrency } from "@app/types/currency";
 import type { BillingPeriod } from "@app/types/plan";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsAdmin } from "@front-api/middlewares/ensure_is_admin";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
