+++ front-api/routes/w/[wId]/webhook_sources/service-data.ts
@@ -9,7 +9,7 @@ import type {
 } from "@app/types/triggers/webhooks";
 import { isWebhookProvider } from "@app/types/triggers/webhooks";
 import { workspaceApp } from "@front-api/middlewares/ctx";
-import { ensureIsAdmin } from "@front-api/middlewares/ensure_is_admin";
+import { ensureIsAdmin } from "@front-api/middlewares/ensure_role";
 import { apiError, type HandlerResult } from "@front-api/middlewares/utils";
 import { validate } from "@front-api/middlewares/validator";
 import { z } from "zod";
